Fixed Drag and scrolling at same time.
