+++ lib/multi_image_picker_view.dart
@@ -2,5 +2,6 @@ library multi_image_picker_view;
 
 export './src/multi_image_picker_view.dart';
 export './src/multi_image_picker_controller.dart';
+export './src/multi_image_picker_components.dart';
 export './src/image_file.dart';
 export './src/image_file_view/image_file_view.dart';
