+++ pubspec.yaml
@@ -13,7 +13,8 @@ dependencies:
   flutter:
     sdk: flutter
 
-  flutter_reorderable_grid_view: ^3.1.3
+  flutter_reorderable_grid_view: ^4.0.0
+  reorderable_grid_view: ^2.2.6
   file_picker: ^5.0.1
 
 dev_dependencies:
